refactor: decompose main.rs into focused binary-local modules
Split the sandbox binary's 2,776-line main.rs along the seams identified
in issue #113. The existing `impl App { ... }` methods move to new
binary-local modules while the `impl ApplicationHandler for App` (window
event loop) stays in main.rs for now.

New modules:
- `app`: `App` state, `PendingUpload`, `GenTransition`, simulation-rate
  constants, and the core init/logging methods (`new`, `log_msg`,
  `print_session_summary`).
- `gpu_orchestration`: `ensure_gpu_kernel` — fused-kernel lifecycle and
  upload staging.
- `evolution`: evolution action handling plus the async generation
  transition state machine (`handle_evolution_action`,
  `spawn_evolution_population`, `spawn_population_from_configs`,
  `advance_generation`, `poll_gen_transition`, `try_finish_generation`,
  `spawn_agent`, `spawn_child`).
- `snapshot`: `record_agent_histories` helper for sparkline histories.
- `render_pipeline`: `build_hud_bars`, `pick_agent_at_cursor`.

App fields are raised to `pub(crate)` so sibling modules can share the
state. No behavioural change; all 122 sandbox tests still pass.

Refs #113
