+++ crates/xagent-sandbox/src/main.rs
@@ -1,18 +1,21 @@
-use std::collections::VecDeque;
+mod app;
+mod evolution;
+mod gpu_orchestration;
+mod render_pipeline;
+mod snapshot;
+
 use std::sync::Arc;
-use std::time::{Duration, Instant};
+use std::time::Instant;
 
 use clap::Parser;
-use egui_wgpu;
 use glam::Vec3;
 use log::info;
-use rand::Rng;
 
 use winit::application::ApplicationHandler;
 use winit::event::{ElementState, MouseButton, MouseScrollDelta, WindowEvent};
 use winit::event_loop::{ActiveEventLoop, EventLoop};
 use winit::keyboard::{KeyCode, PhysicalKey};
-use winit::window::{Window, WindowAttributes, WindowId};
+use winit::window::{WindowAttributes, WindowId};
 use xagent_shared::{BrainConfig, FullConfig, GovernorConfig, WorldConfig};
 
 use xagent_brain::buffers::{
@@ -22,21 +25,19 @@ use xagent_brain::buffers::{
     P_MOTOR_FWD_OUT, P_MOTOR_TURN_OUT, P_POS_X, P_POS_Y, P_POS_Z, P_PREDICTION_ERROR,
     P_TICKS_ALIVE, P_URGENCY_OUT, P_VEL_X, P_VEL_Y, P_VEL_Z, P_YAW,
 };
-use xagent_brain::{AgentBrainState, GpuKernel};
-use xagent_sandbox::agent::{mutate_brain_state, mutate_config, Agent, MAX_AGENTS};
-use xagent_sandbox::governor::{check_existing_session, reset_database, AdvanceResult, Governor};
+use xagent_sandbox::agent::{mutate_config, MAX_AGENTS};
 use xagent_sandbox::headless;
 use xagent_sandbox::overlay;
-use xagent_sandbox::renderer::camera::Camera;
 use xagent_sandbox::renderer::font::TextItem;
-use xagent_sandbox::renderer::hud::HudBar;
 use xagent_sandbox::renderer::{GpuMesh, InstanceData, Renderer};
 use xagent_sandbox::ui::{
-    AgentSnapshot, EguiIntegration, EvolutionAction, EvolutionSnapshot, EvolutionState,
-    ReplayState, SortMode, Tab, TabContext, WorldSnapshot,
+    AgentSnapshot, EguiIntegration, EvolutionAction, EvolutionState, SortMode, Tab, TabContext,
 };
 
 use xagent_sandbox::world::WorldState;
+
+use crate::app::{App, REBUILD_THROTTLE, SIM_DT, SIM_RATE};
+use crate::snapshot::record_agent_histories;
 
 // ── CLI ────────────────────────────────────────────────────────────────
 
@@ -188,1032 +189,6 @@ fn print_config(config: &FullConfig) {
         config.world.seed,
     );
     println!("──────────────────────────────────────────────────");
-}
-
-/// Data collected on the main thread for upload once the background-created
-/// fused kernel is ready.  Avoids passing world/agent refs across threads.
-struct PendingUpload {
-    heights: Vec<f32>,
-    biomes: Vec<u32>,
-    food_pos: Vec<(f32, f32, f32)>,
-    food_consumed: Vec<bool>,
-    food_timers: Vec<f32>,
-    agent_data: Vec<(glam::Vec3, f32, f32, usize, usize)>,
-}
-
-/// Base simulation rate in Hz. The simulation runs at this rate (scaled by
-/// speed multiplier) regardless of rendering frame rate.
-const SIM_RATE: f32 = 60.0;
-const SIM_DT: f32 = 1.0 / SIM_RATE;
-
-/// Minimum interval between expensive mesh rebuilds and snapshot copies.
-/// At 100 ms (10 Hz) the visual difference is indistinguishable from per-frame.
-const REBUILD_THROTTLE: Duration = Duration::from_millis(100);
-
-/// Record per-tick telemetry into agent sparkline histories.
-fn record_agent_histories(agent: &mut Agent) {
-    let cap = 10_000;
-    macro_rules! push_hist {
-        ($h:expr, $v:expr) => {
-            if $h.len() >= cap {
-                $h.pop_front();
-            }
-            $h.push_back($v);
-        };
-    }
-    push_hist!(
-        agent.prediction_error_history,
-        agent.cached_prediction_error
-    );
-    push_hist!(
-        agent.exploration_rate_history,
-        agent.cached_exploration_rate
-    );
-    let ef = agent.body.body.internal.energy / agent.body.body.internal.max_energy.max(0.001);
-    push_hist!(agent.energy_history, ef.clamp(0.0, 1.0));
-    let inf =
-        agent.body.body.internal.integrity / agent.body.body.internal.max_integrity.max(0.001);
-    push_hist!(agent.integrity_history, inf.clamp(0.0, 1.0));
-    push_hist!(agent.fatigue_history, agent.cached_fatigue_factor);
-}
-
-/// State machine for non-blocking generation transitions.
-/// Each variant represents one phase of the multi-frame handoff.
-enum GenTransition {
-    /// Waiting for async readback of the best agent's brain state.
-    AwaitingReadback { result: AdvanceResult },
-    /// Readback collected; waiting for staging buffers to drain before reset.
-    AwaitingReset {
-        result: AdvanceResult,
-        inherited_state: Option<AgentBrainState>,
-        /// Whether population has already been spawned (prevents re-spawn on retry).
-        spawned: bool,
-    },
-}
-
-struct App {
-    brain_config: BrainConfig,
-    world_config: WorldConfig,
-
-    renderer: Option<Renderer>,
-    window: Option<Arc<Window>>,
-    camera: Camera,
-
-    // Static terrain mesh (built once)
-    terrain_gpu: Option<GpuMesh>,
-
-    // Simulation
-    world: Option<WorldState>,
-    agents: Vec<Agent>,
-    next_agent_id: u32,
-    tick: u64,
-    food_dirty: bool,
-    food_gpu: Option<GpuMesh>,
-
-    last_frame: Instant,
-
-    /// Fixed-timestep accumulator — simulation ticks run at SIM_RATE Hz,
-    /// decoupled from the render frame rate.
-    sim_accumulator: f64,
-
-    // Speed controls (multiplier for simulation rate)
-    speed_multiplier: u32,
-    paused: bool,
-    render_3d: bool,
-
-    // Telemetry selection
-    selected_agent_idx: usize,
-    viewport_hovered: bool,
-    chart_window: usize,
-    orbit_mode: bool,
-
-    // GPU instancing for agents
-    agent_instance_buffer: Option<wgpu::Buffer>,
-    agent_instance_count: u32,
-
-    // FPS tracking
-    frame_times: VecDeque<Instant>,
-    fps: f32,
-
-    // Cached HUD data — only rebuilt when a simulation tick runs
-    cached_hud_bars: Vec<HudBar>,
-    hud_dirty: bool,
-
-    // Cursor position for click-to-select
-    cursor_pos: (f64, f64),
-
-    // Heatmap overlay
-    heatmap_enabled: bool,
-    heatmap_gpu: Option<GpuMesh>,
-    heatmap_dirty: bool,
-
-    // Trail overlay for selected agent
-    trail_gpu: Option<GpuMesh>,
-
-    // Throttle timers — limit expensive rebuilds to ~10 Hz
-    last_mesh_rebuild: Instant,
-    last_snapshot_rebuild: Instant,
-    snap_dirty: bool,
-    cached_agent_snaps: Vec<AgentSnapshot>,
-    last_snap_chart_window: usize,
-
-    // Selection marker above focused agent
-    marker_gpu: Option<GpuMesh>,
-
-    // egui integration (IDE-style UI overlay)
-    egui: Option<EguiIntegration>,
-
-    // Console log buffer for the bottom panel
-    console_log: VecDeque<String>,
-
-    // Evolution governor (created on Start/Resume, not on App init)
-    governor: Option<Governor>,
-    evo_snapshot: EvolutionSnapshot,
-    evo_wall_accumulated: f64,
-    evo_wall_segment_start: Option<Instant>,
-    tps_tick_count: u64,
-    tps_last_reset: Instant,
-    tps_display: f64,
-    db_path: String,
-    governor_config: GovernorConfig,
-
-    // Non-blocking generation transition state machine
-    gen_transition: Option<GenTransition>,
-
-    // GPU fused kernel (all simulation computation in single dispatch)
-    gpu_kernel: Option<GpuKernel>,
-    pending_kernel: Option<std::thread::JoinHandle<GpuKernel>>,
-    /// Data collected for upload once the background kernel is ready.
-    pending_upload: Option<PendingUpload>,
-    /// Inherited brain state + mutation strength deferred until a background
-    /// kernel finishes creation (when population size changes at generation
-    /// boundary and `can_reuse` is false).
-    deferred_inherited: Option<(AgentBrainState, f32)>,
-
-    // egui_dock tab state
-    dock_state: egui_dock::DockState<Tab>,
-
-    // Adaptive GPU tick budget — keeps dispatch under ~8ms wall time
-    gpu_tick_budget: u32,
-
-    // Diagnostic: log first readback Y vs terrain height
-    readback_logged: bool,
-
-    // Sidebar sort mode
-    sort_mode: SortMode,
-
-    // World snapshot for mini-map UI
-    world_snapshot: WorldSnapshot,
-
-    // Replay recording
-    recording: Option<xagent_sandbox::replay::GenerationRecording>,
-    last_recording: Option<xagent_sandbox::replay::GenerationRecording>,
-
-    // Replay playback state
-    replay_state: ReplayState,
-}
-
-impl App {
-    fn new(
-        brain_config: BrainConfig,
-        world_config: WorldConfig,
-        governor_config: GovernorConfig,
-        db_path: &str,
-    ) -> Self {
-        // Check for existing session to determine initial UI state
-        let mut evo_snapshot = EvolutionSnapshot::default();
-        evo_snapshot.edit_brain = brain_config.clone();
-        evo_snapshot.edit_governor = governor_config.clone();
-
-        match check_existing_session(db_path) {
-            Some((gen, prev_gov, prev_brain)) => {
-                evo_snapshot.state = EvolutionState::HasSession { generation: gen };
-                evo_snapshot.generation = gen;
-                evo_snapshot.current_config = Some(prev_brain);
-                evo_snapshot.population_size = prev_gov.population_size;
-                evo_snapshot.tick_budget = prev_gov.tick_budget;
-                evo_snapshot.elitism_count = prev_gov.elitism_count;
-                evo_snapshot.patience = prev_gov.patience;
-                evo_snapshot.max_generations = prev_gov.max_generations;
-                evo_snapshot.eval_repeats = prev_gov.eval_repeats;
-                evo_snapshot.num_islands = prev_gov.num_islands;
-                evo_snapshot.migration_interval = prev_gov.migration_interval;
-                // Also load tree/fitness from DB for the summary view
-                if let Ok(mut gov) = Governor::resume(db_path) {
-                    evo_snapshot.tree_nodes = gov.tree_nodes();
-                    evo_snapshot.current_node_id = gov.current_node_id;
-                    evo_snapshot
-                        .fitness_history
-                        .clone_from(gov.fitness_history_by_island());
-                    evo_snapshot.best_fitness = gov.best_score();
-                }
-                println!(
-                    "[GOVERNOR] Found previous session at generation {} in {}",
-                    gen, db_path
-                );
-            }
-            None => {
-                evo_snapshot.state = EvolutionState::Idle;
-                println!("[GOVERNOR] No previous session — ready to configure");
-            }
-        }
-
-        Self {
-            brain_config,
-            world_config,
-            renderer: None,
-            window: None,
-            camera: Camera::new(16.0 / 9.0),
-            terrain_gpu: None,
-            world: None,
-            agents: Vec::new(),
-            next_agent_id: 0,
-            tick: 0,
-            food_dirty: true,
-            food_gpu: None,
-            last_frame: Instant::now(),
-            sim_accumulator: 0.0,
-            speed_multiplier: 1,
-            paused: true,
-            render_3d: true,
-            selected_agent_idx: 0,
-            viewport_hovered: false,
-            chart_window: 120,
-            orbit_mode: false,
-            readback_logged: false,
-            agent_instance_buffer: None,
-            agent_instance_count: 0,
-            frame_times: VecDeque::new(),
-            fps: 0.0,
-            cached_hud_bars: Vec::new(),
-            hud_dirty: true,
-            cursor_pos: (0.0, 0.0),
-            heatmap_enabled: false,
-            heatmap_gpu: None,
-            heatmap_dirty: false,
-            trail_gpu: None,
-            last_mesh_rebuild: Instant::now(),
-            last_snapshot_rebuild: Instant::now(),
-            snap_dirty: true,
-            cached_agent_snaps: Vec::new(),
-            last_snap_chart_window: 120,
-            marker_gpu: None,
-            egui: None,
-            console_log: VecDeque::new(),
-            governor: None,
-            evo_snapshot,
-            evo_wall_accumulated: 0.0,
-            evo_wall_segment_start: None,
-            tps_tick_count: 0,
-            tps_last_reset: Instant::now(),
-            tps_display: 0.0,
-            db_path: db_path.to_string(),
-            governor_config,
-            gen_transition: None,
-            gpu_kernel: None,
-            pending_kernel: None,
-            pending_upload: None,
-            deferred_inherited: None,
-            dock_state: egui_dock::DockState::new(vec![Tab::Evolution, Tab::Sandbox]),
-            gpu_tick_budget: 32,
-            sort_mode: SortMode::Id,
-            world_snapshot: WorldSnapshot::default(),
-            recording: None,
-            last_recording: None,
-            replay_state: ReplayState::default(),
-        }
-    }
-
-    /// Spawn a new agent with the given BrainConfig at a safe random position
-    /// (not in a danger biome). The brain_idx is the agent's array index.
-    fn spawn_agent(&mut self, config: BrainConfig, generation: u32) {
-        if self.agents.len() >= MAX_AGENTS {
-            return;
-        }
-        let Some(world) = &self.world else { return };
-
-        let pos = world.safe_spawn_position();
-
-        let id = self.next_agent_id;
-        self.next_agent_id += 1;
-
-        let brain_idx = self.agents.len() as u32;
-        let mut agent = Agent::new(id, pos, brain_idx, config, self.tick);
-        agent.generation = generation;
-
-        self.agents.push(agent);
-    }
-
-    /// Ensure GpuKernel is initialized for the current population.
-    fn ensure_gpu_kernel(&mut self) {
-        if self.gpu_kernel.is_some() {
-            return;
-        }
-        if self.agents.is_empty() {
-            return;
-        }
-        let world = match &self.world {
-            Some(w) => w,
-            None => return,
-        };
-
-        // Check if background creation finished.
-        if let Some(ref handle) = self.pending_kernel {
-            if handle.is_finished() {
-                let handle = self.pending_kernel.take().unwrap();
-                let mk = handle
-                    .join()
-                    .expect("fused kernel background thread panicked");
-                // Upload world + agent data (fast, main thread).
-                if let Some(upload) = self.pending_upload.take() {
-                    mk.upload_world(
-                        &upload.heights,
-                        &upload.biomes,
-                        &upload.food_pos,
-                        &upload.food_consumed,
-                        &upload.food_timers,
-                    );
-                    mk.upload_agents(&upload.agent_data);
-                }
-                let ac = mk.agent_count();
-
-                // Apply deferred inherited state from a generation transition
-                // that occurred while the kernel was being recreated.
-                if let Some((ref state, mutation_strength)) = self.deferred_inherited.take() {
-                    let repeats = self.governor_config.eval_repeats.max(1);
-                    let n = self.agents.len();
-                    let champion = state.clone();
-                    mk.batch_write_agent_states(n, |i| {
-                        if i < repeats {
-                            champion.clone()
-                        } else {
-                            mutate_brain_state(state, mutation_strength)
-                        }
-                    });
-                }
-
-                self.gpu_kernel = Some(mk);
-                self.log_msg(format!("[GPU] GpuKernel ready ({} agents)", ac));
-            }
-            return; // still creating
-        }
-
-        // Collect data for upload (kept on main thread).
-        let agent_count = self.agents.len() as u32;
-        let food_count = world.food_items.len();
-        let brain_config = self.brain_config.clone();
-        let world_config = world.config.clone();
-
-        self.pending_upload = Some(PendingUpload {
-            heights: world.terrain.heights.clone(),
-            biomes: world.biome_map.grid_as_u32(),
-            food_pos: world
-                .food_items
-                .iter()
-                .map(|f| (f.position.x, f.position.y, f.position.z))
-                .collect(),
-            food_consumed: world.food_items.iter().map(|f| f.consumed).collect(),
-            food_timers: world.food_items.iter().map(|f| f.respawn_timer).collect(),
-            agent_data: self
-                .agents
-                .iter()
-                .map(|a| {
-                    (
-                        a.body.body.position,
-                        a.body.body.internal.max_energy,
-                        a.body.body.internal.max_integrity,
-                        a.brain_config.memory_capacity,
-                        a.brain_config.processing_slots,
-                    )
-                })
-                .collect(),
-        });
-
-        // Spawn background thread for device + shader compilation.
-        self.pending_kernel = Some(std::thread::spawn(move || {
-            GpuKernel::new(agent_count, food_count, &brain_config, &world_config)
-        }));
-        self.log_msg("[GPU] Creating GpuKernel (background)...".into());
-    }
-
-    fn handle_evolution_action(&mut self, action: EvolutionAction) {
-        match action {
-            EvolutionAction::None => {}
-            EvolutionAction::Start => {
-                let brain_config = self.evo_snapshot.edit_brain.clone();
-                let gov_config = self.evo_snapshot.edit_governor.clone();
-                let world_json = serde_json::to_string(&self.world_config).unwrap_or_default();
-                match Governor::new(
-                    &self.db_path,
-                    gov_config.clone(),
-                    &brain_config,
-                    &world_json,
-                ) {
-                    Ok(gov) => {
-                        self.governor = Some(gov);
-                        self.governor_config = gov_config.clone();
-                        self.brain_config = brain_config;
-                        self.evo_snapshot.state = EvolutionState::Running;
-                        self.evo_snapshot.population_size = gov_config.population_size;
-                        self.evo_snapshot.tick_budget = gov_config.tick_budget;
-                        self.evo_snapshot.elitism_count = gov_config.elitism_count;
-                        self.evo_snapshot.patience = gov_config.patience;
-                        self.evo_snapshot.max_generations = gov_config.max_generations;
-                        self.evo_snapshot.eval_repeats = gov_config.eval_repeats;
-                        self.evo_snapshot.num_islands = gov_config.num_islands;
-                        self.evo_snapshot.migration_interval = gov_config.migration_interval;
-                        self.evo_wall_accumulated = 0.0;
-                        self.evo_wall_segment_start = Some(Instant::now());
-                        self.tps_tick_count = 0;
-                        self.tps_last_reset = Instant::now();
-                        self.tps_display = 0.0;
-                        self.paused = false;
-                        self.gpu_kernel = None;
-                        self.pending_kernel = None;
-                        self.pending_upload = None;
-                        self.spawn_evolution_population();
-                        self.ensure_gpu_kernel();
-                        self.log_msg("[EVOLUTION] Started new run".into());
-                    }
-                    Err(e) => {
-                        self.log_msg(format!("[EVOLUTION] Failed to start: {}", e));
-                    }
-                }
-            }
-            EvolutionAction::Resume => match Governor::resume(&self.db_path) {
-                Ok(gov) => {
-                    let cfg = gov.current_config();
-                    self.evo_snapshot.state = EvolutionState::Running;
-                    self.evo_snapshot.generation = gov.generation;
-                    self.governor_config = gov.config.clone();
-                    if let Some(c) = &cfg {
-                        self.brain_config = c.clone();
-                    }
-                    self.governor = Some(gov);
-                    self.evo_wall_accumulated = 0.0;
-                    self.evo_wall_segment_start = Some(Instant::now());
-                    self.tps_tick_count = 0;
-                    self.tps_last_reset = Instant::now();
-                    self.tps_display = 0.0;
-                    self.paused = false;
-                    self.gpu_kernel = None;
-                    self.pending_kernel = None;
-                    self.pending_upload = None;
-                    self.spawn_evolution_population();
-                    self.ensure_gpu_kernel();
-                    self.log_msg("[EVOLUTION] Resumed from database".into());
-                }
-                Err(e) => {
-                    self.log_msg(format!("[EVOLUTION] Failed to resume: {}", e));
-                }
-            },
-            EvolutionAction::Pause => {
-                self.evo_snapshot.state = EvolutionState::Paused;
-                self.paused = true;
-                self.snap_dirty = true;
-                if let Some(start) = self.evo_wall_segment_start.take() {
-                    self.evo_wall_accumulated += start.elapsed().as_secs_f64();
-                }
-                self.tps_display = 0.0;
-                self.log_msg("[EVOLUTION] Paused".into());
-            }
-            EvolutionAction::Unpause => {
-                self.evo_snapshot.state = EvolutionState::Running;
-                self.paused = false;
-                self.evo_wall_segment_start = Some(Instant::now());
-                self.tps_tick_count = 0;
-                self.tps_last_reset = Instant::now();
-                self.log_msg("[EVOLUTION] Resumed".into());
-            }
-            EvolutionAction::Reset => {
-                self.governor = None;
-                self.gpu_kernel = None;
-                self.pending_kernel = None;
-                self.pending_upload = None;
-                self.agents.clear();
-                self.next_agent_id = 0;
-                self.tick = 0;
-                self.paused = true;
-                if let Err(e) = reset_database(&self.db_path) {
-                    self.log_msg(format!("[EVOLUTION] Failed to reset DB: {}", e));
-                }
-                self.evo_snapshot = EvolutionSnapshot::default();
-                self.evo_snapshot.edit_brain = self.brain_config.clone();
-                self.evo_snapshot.edit_governor = self.governor_config.clone();
-                self.log_msg("[EVOLUTION] Reset — ready to start fresh".into());
-            }
-        }
-    }
-
-    fn spawn_evolution_population(&mut self) {
-        self.agents.clear();
-        self.next_agent_id = 0;
-        self.tick = 0;
-        let seed = if let Some(gov) = &self.governor {
-            gov.current_config().unwrap_or(self.brain_config.clone())
-        } else {
-            self.brain_config.clone()
-        };
-        let pop_size = self.governor_config.population_size;
-        let repeats = self.governor_config.eval_repeats.max(1);
-        let unique_count = (pop_size / repeats).max(1);
-
-        // Build unique configs matching breed_next_generation structure
-        // so that reduce_fitness grouping (agent_index / eval_repeats)
-        // correctly averages same-config runs.
-        let mut unique_configs = vec![seed.clone()]; // slot 0: champion
-        for _ in 1..unique_count {
-            unique_configs.push(mutate_config(&seed));
-        }
-
-        // Repeat each config eval_repeats times
-        for uc in &unique_configs {
-            for _ in 0..repeats {
-                if self.agents.len() >= pop_size {
-                    break;
-                }
-                self.spawn_agent(uc.clone(), 0);
-            }
-        }
-
-        // Start replay recording
-        if let Some(world) = &self.world {
-            let agent_info: Vec<(u32, [f32; 3])> =
-                self.agents.iter().map(|a| (a.id, a.color)).collect();
-            let initial_food: Vec<[f32; 3]> = world
-                .food_items
-                .iter()
-                .map(|f| [f.position.x, f.position.y, f.position.z])
-                .collect();
-            let gen = self.governor.as_ref().map_or(0, |g| g.generation as u32);
-            self.recording = Some(xagent_sandbox::replay::GenerationRecording::new(
-                gen,
-                &agent_info,
-                &initial_food,
-                self.governor_config.tick_budget as usize,
-                self.brain_config.vision_width,
-                self.brain_config.vision_height,
-            ));
-        }
-    }
-
-    fn spawn_population_from_configs(&mut self, configs: &[BrainConfig]) {
-        self.agents.clear();
-        self.next_agent_id = 0;
-        self.tick = 0;
-        for cfg in configs {
-            self.spawn_agent(cfg.clone(), 0);
-        }
-
-        // Start replay recording
-        if let Some(world) = &self.world {
-            let agent_info: Vec<(u32, [f32; 3])> =
-                self.agents.iter().map(|a| (a.id, a.color)).collect();
-            let initial_food: Vec<[f32; 3]> = world
-                .food_items
-                .iter()
-                .map(|f| [f.position.x, f.position.y, f.position.z])
-                .collect();
-            let gen = self.governor.as_ref().map_or(0, |g| g.generation as u32);
-            self.recording = Some(xagent_sandbox::replay::GenerationRecording::new(
-                gen,
-                &agent_info,
-                &initial_food,
-                self.governor_config.tick_budget as usize,
-                self.brain_config.vision_width,
-                self.brain_config.vision_height,
-            ));
-        }
-    }
-
-    /// Evaluate the current generation and begin the async transition.
-    /// The actual GPU work (readback, reset, upload) spans multiple frames
-    /// via `poll_gen_transition`.
-    fn advance_generation(&mut self) {
-        // Persist the recording to SQLite before moving to the next generation
-        if let (Some(ref recording), Some(ref mut gov)) = (&self.recording, &mut self.governor) {
-            gov.store_recording(recording);
-        }
-        self.last_recording = self.recording.take();
-        let wall_secs = self.evo_wall_accumulated
-            + self
-                .evo_wall_segment_start
-                .map(|s| s.elapsed().as_secs_f64())
-                .unwrap_or(0.0);
-
-        // Evaluate fitness and decide whether evolution continues.
-        // Only kick off the async brain-state readback when the result is
-        // Continue — when Finished, no readback is needed.
-        let (result, readback_requested) = {
-            let gov = match self.governor.as_mut() {
-                Some(g) => g,
-                None => return,
-            };
-
-            let fitness = gov.evaluate(&self.agents);
-            gov.log_generation(&fitness);
-            gov.update_wall_time(wall_secs);
-
-            let result = gov.advance(&fitness);
-
-            // Only read back the best agent's brain state when evolution
-            // continues — Finished needs no inherited state.
-            let mut readback_requested = false;
-            if matches!(result, AdvanceResult::Continue { .. }) {
-                let best_idx = fitness.first().map(|f| f.agent_index).unwrap_or(0);
-                if let Some(a) = self.agents.get(best_idx) {
-                    if let Some(mk) = self.gpu_kernel.as_mut() {
-                        if mk.request_agent_state(a.brain_idx) {
-                            readback_requested = true;
-                        }
-                    }
-                }
-            }
-
-            (result, readback_requested)
-        };
-
-        match result {
-            AdvanceResult::Continue { .. } if readback_requested => {
-                self.gen_transition = Some(GenTransition::AwaitingReadback { result });
-            }
-            AdvanceResult::Continue { .. } => {
-                // No readback was issued (no best agent or no GPU kernel) —
-                // skip directly to reset to avoid deadlocking in AwaitingReadback.
-                self.gen_transition = Some(GenTransition::AwaitingReset {
-                    result,
-                    inherited_state: None,
-                    spawned: false,
-                });
-            }
-            AdvanceResult::Finished { .. } => {
-                // Skip readback phase entirely — go straight to finish.
-                self.gen_transition = Some(GenTransition::AwaitingReset {
-                    result,
-                    inherited_state: None,
-                    spawned: false,
-                });
-            }
-        }
-    }
-
-    /// Drive the multi-frame generation transition state machine.
-    /// Called every frame; returns true while a transition is still in progress.
-    fn poll_gen_transition(&mut self) -> bool {
-        let transition = match self.gen_transition.take() {
-            Some(t) => t,
-            None => return false,
-        };
-
-        match transition {
-            GenTransition::AwaitingReadback { result } => {
-                if self.gpu_kernel.is_none() {
-                    // No kernel — proceed without inherited state.
-                    self.gen_transition = Some(GenTransition::AwaitingReset {
-                        result,
-                        inherited_state: None,
-                        spawned: false,
-                    });
-                    return true;
-                }
-
-                // Poll for the async agent state readback.
-                let collected = self.gpu_kernel.as_mut().unwrap().try_collect_agent_state();
-
-                match collected {
-                    Some(Some(state)) => {
-                        // Readback complete — move to reset phase.
-                        self.gen_transition = Some(GenTransition::AwaitingReset {
-                            result,
-                            inherited_state: Some(state),
-                            spawned: false,
-                        });
-                    }
-                    Some(None) => {
-                        // map_async error — proceed without inherited state
-                        // rather than hanging in AwaitingReadback forever.
-                        self.gen_transition = Some(GenTransition::AwaitingReset {
-                            result,
-                            inherited_state: None,
-                            spawned: false,
-                        });
-                    }
-                    None => {
-                        // Still waiting — re-enqueue for next frame.
-                        self.gen_transition = Some(GenTransition::AwaitingReadback { result });
-                    }
-                }
-                true
-            }
-            GenTransition::AwaitingReset {
-                result,
-                inherited_state,
-                spawned,
-            } => {
-                let (done, spawned) =
-                    self.try_finish_generation(&result, inherited_state.as_ref(), spawned);
-                if !done {
-                    // Reset not ready yet — re-enqueue for next frame.
-                    self.gen_transition = Some(GenTransition::AwaitingReset {
-                        result,
-                        inherited_state,
-                        spawned,
-                    });
-                    return true;
-                }
-                false
-            }
-        }
-    }
-
-    /// Try to complete a generation transition. Returns `(done, spawned)`
-    /// where `done` is false if the GPU staging buffers aren't ready yet
-    /// (caller should retry next frame), and `spawned` tracks whether
-    /// population has been spawned (to avoid re-spawning on retry).
-    fn try_finish_generation(
-        &mut self,
-        result: &AdvanceResult,
-        inherited_state: Option<&AgentBrainState>,
-        mut spawned: bool,
-    ) -> (bool, bool) {
-        match result {
-            AdvanceResult::Continue {
-                configs,
-                messages,
-                mutation_strength,
-            } => {
-                if !spawned {
-                    for msg in messages {
-                        self.log_msg(msg.clone());
-                    }
-                    self.spawn_population_from_configs(configs);
-                    spawned = true;
-                }
-                let repeats = self.governor_config.eval_repeats.max(1);
-
-                // Reuse existing kernel if population size matches,
-                // otherwise drop and recreate in background.
-                let can_reuse = self
-                    .gpu_kernel
-                    .as_ref()
-                    .is_some_and(|mk| mk.agent_count() == self.agents.len() as u32);
-
-                if can_reuse {
-                    let mk = self.gpu_kernel.as_mut().unwrap();
-                    // Non-blocking reset — return false to retry next frame
-                    // instead of falling back to a blocking busy-wait.
-                    if !mk.try_reset_agents(&self.brain_config) {
-                        return (false, spawned);
-                    }
-                    let agent_data: Vec<_> = self
-                        .agents
-                        .iter()
-                        .map(|a| {
-                            (
-                                a.body.body.position,
-                                a.body.body.internal.max_energy,
-                                a.body.body.internal.max_integrity,
-                                a.brain_config.memory_capacity,
-                                a.brain_config.processing_slots,
-                            )
-                        })
-                        .collect();
-                    mk.upload_agents(&agent_data);
-                } else {
-                    self.gpu_kernel = None;
-                    self.pending_kernel = None;
-                    self.pending_upload = None;
-                    self.ensure_gpu_kernel();
-                    // Kernel is being recreated in the background —
-                    // keep the transition active until it's ready.
-                    if self.gpu_kernel.is_none() {
-                        return (false, spawned);
-                    }
-                }
-
-                // Inherit learned weights: champions get exact brain state,
-                // mutants get perturbed weights for neuroevolution.
-                if let Some(state) = inherited_state {
-                    if let Some(ref mk) = self.gpu_kernel {
-                        let ms = *mutation_strength;
-                        let n = self.agents.len();
-                        // Pre-clone once for all champion slots to avoid
-                        // cloning inside the hot closure on every call.
-                        let champion = state.clone();
-                        mk.batch_write_agent_states(n, |i| {
-                            if i < repeats {
-                                champion.clone()
-                            } else {
-                                mutate_brain_state(state, ms)
-                            }
-                        });
-                    } else {
-                        // Kernel is being recreated in the background
-                        // (population size changed). Stash inherited state
-                        // so it can be applied once the kernel is ready.
-                        self.deferred_inherited = Some((state.clone(), *mutation_strength));
-                    }
-                }
-            }
-            AdvanceResult::Finished { messages } => {
-                for msg in messages {
-                    self.log_msg(msg.clone());
-                }
-                self.evo_snapshot.state = EvolutionState::Paused;
-                self.paused = true;
-            }
-        }
-        (true, spawned)
-    }
-
-    fn log_msg(&mut self, message: String) {
-        println!("{}", message);
-        self.console_log.push_back(message);
-        if self.console_log.len() > 200 {
-            self.console_log.pop_front();
-        }
-    }
-
-    /// Spawn a child agent near a parent, with mutated config.
-    #[allow(dead_code)]
-    fn spawn_child(&mut self, parent_idx: usize) {
-        if self.agents.len() >= MAX_AGENTS {
-            return;
-        }
-        let parent = &self.agents[parent_idx];
-        let parent_id = parent.id;
-        let parent_gen = parent.generation;
-        let parent_pos = parent.body.body.position;
-        let parent_config = parent.brain_config.clone();
-
-        let child_config = mutate_config(&parent_config);
-        let mut rng = rand::rng();
-        let offset = Vec3::new(
-            rng.random_range(-5.0..5.0_f32),
-            0.0,
-            rng.random_range(-5.0..5.0_f32),
-        );
-        let child_pos = parent_pos + offset;
-
-        let id = self.next_agent_id;
-        self.next_agent_id += 1;
-
-        let Some(world) = &self.world else { return };
-        let half = world.config.world_size / 2.0 - 1.0;
-        let cx = child_pos.x.clamp(-half, half);
-        let cz = child_pos.z.clamp(-half, half);
-        let cy = world.terrain.height_at(cx, cz) + 1.0;
-
-        let brain_idx = self.agents.len() as u32;
-        let mut child = Agent::new(
-            id,
-            Vec3::new(cx, cy, cz),
-            brain_idx,
-            child_config,
-            self.tick,
-        );
-        child.generation = parent_gen + 1;
-
-        println!(
-            "[REPRODUCE] Agent {} (gen {}) → child Agent {} (gen {}) at ({:.1}, {:.1})",
-            parent_id, parent_gen, id, child.generation, cx, cz
-        );
-        // `mem_cost` / `proc_cost`: metabolic-cost proxies — kernel widths
-        // are fixed at MEMORY_CAP=128, RECALL_K=16 (see issue #106).
-        println!(
-            "  Child config: mem_cost={} proc_cost={} dim={} lr={:.4} decay={:.4}",
-            child.brain_config.memory_capacity,
-            child.brain_config.processing_slots,
-            child.brain_config.representation_dimension,
-            child.brain_config.learning_rate,
-            child.brain_config.decay_rate,
-        );
-
-        self.agents.push(child);
-    }
-
-    /// Pick the agent closest to the cursor via screen-space projection.
-    fn pick_agent_at_cursor(&mut self) {
-        let Some(renderer) = &self.renderer else {
-            return;
-        };
-        let w = renderer.config.width as f32;
-        let h = renderer.config.height as f32;
-        if w < 1.0 || h < 1.0 {
-            return;
-        }
-
-        // Normalized device coordinates [-1, 1]
-        let ndc_x = (self.cursor_pos.0 as f32 / w) * 2.0 - 1.0;
-        let ndc_y = 1.0 - (self.cursor_pos.1 as f32 / h) * 2.0;
-
-        let vp = self.camera.view_projection_matrix();
-        let mut best_idx: Option<usize> = None;
-        let mut best_dist_sq = f32::MAX;
-
-        for (i, agent) in self.agents.iter().enumerate() {
-            if !agent.body.body.alive {
-                continue;
-            }
-            let pos = agent.body.body.position;
-            let clip = vp * glam::Vec4::new(pos.x, pos.y, pos.z, 1.0);
-            if clip.w <= 0.0 {
-                continue;
-            } // behind camera
-            let sx = clip.x / clip.w;
-            let sy = clip.y / clip.w;
-            let d = (sx - ndc_x).powi(2) + (sy - ndc_y).powi(2);
-            if d < best_dist_sq {
-                best_dist_sq = d;
-                best_idx = Some(i);
-            }
-        }
-
-        // Only select if click is reasonably close (within ~10% of screen)
-        if let Some(idx) = best_idx {
-            if best_dist_sq < 0.05 {
-                self.selected_agent_idx = idx;
-                self.agents[idx].trail_dirty = true;
-                let a = &self.agents[idx];
-                println!("[SELECT] Agent {} (gen {})", a.id, a.generation);
-                self.hud_dirty = true;
-            }
-        }
-    }
-
-    fn print_session_summary(&self) {
-        let total_deaths: u32 = self.agents.iter().map(|a| a.death_count).sum();
-
-        println!();
-        println!("=== xagent Session Summary ===");
-        println!("Total ticks: {}", self.tick);
-        println!("Total agents spawned: {}", self.next_agent_id);
-        println!("Living agents: {}", self.agents.len());
-        println!("Total deaths: {}", total_deaths);
-        println!();
-    }
-
-    /// Build HUD overlay bars for the selected agent.
-    fn build_hud_bars(&self) -> Vec<HudBar> {
-        let Some(agent) = self.agents.get(self.selected_agent_idx) else {
-            return Vec::new();
-        };
-
-        let energy = agent.body.body.internal.energy_signal();
-        let integrity = agent.body.body.internal.integrity_signal();
-        let pred_err = agent.cached_prediction_error;
-        let explore = agent.cached_exploration_rate;
-
-        let bar_w = 0.35;
-        let bar_h = 0.025;
-        let left = -0.98;
-        let top = 0.97;
-        let gap = 0.035;
-        let bg = [0.15, 0.15, 0.15];
-
-        vec![
-            // Energy — green
-            HudBar {
-                x: left,
-                y: top,
-                width: bar_w,
-                height: bar_h,
-                fill: energy,
-                color: [0.2, 0.85, 0.2],
-                bg_color: bg,
-            },
-            // Integrity — blue
-            HudBar {
-                x: left,
-                y: top - gap,
-                width: bar_w,
-                height: bar_h,
-                fill: integrity,
-                color: [0.3, 0.5, 1.0],
-                bg_color: bg,
-            },
-            // Prediction error — red
-            HudBar {
-                x: left,
-                y: top - gap * 2.0,
-                width: bar_w,
-                height: bar_h,
-                fill: pred_err,
-                color: [0.95, 0.2, 0.15],
-                bg_color: bg,
-            },
-            // Exploration rate — yellow
-            HudBar {
-                x: left,
-                y: top - gap * 3.0,
-                width: bar_w,
-                height: bar_h,
-                fill: explore,
-                color: [0.95, 0.85, 0.1],
-                bg_color: bg,
-            },
-        ]
-    }
 }
 
 impl ApplicationHandler for App {
