+++ crates/xagent-sandbox/src/render_pipeline.rs
@@ -1,0 +1,122 @@
+//! Render-side helpers that build UI/overlay geometry from `App` state.
+//!
+//! Hosts `App::build_hud_bars` (HUD overlay bars for the selected agent)
+//! and `App::pick_agent_at_cursor` (screen-space agent picking).  The name
+//! avoids colliding with the library's `renderer` module.
+
+use xagent_sandbox::renderer::hud::HudBar;
+
+use crate::app::App;
+
+impl App {
+    /// Pick the agent closest to the cursor via screen-space projection.
+    pub(crate) fn pick_agent_at_cursor(&mut self) {
+        let Some(renderer) = &self.renderer else {
+            return;
+        };
+        let w = renderer.config.width as f32;
+        let h = renderer.config.height as f32;
+        if w < 1.0 || h < 1.0 {
+            return;
+        }
+
+        // Normalized device coordinates [-1, 1]
+        let ndc_x = (self.cursor_pos.0 as f32 / w) * 2.0 - 1.0;
+        let ndc_y = 1.0 - (self.cursor_pos.1 as f32 / h) * 2.0;
+
+        let vp = self.camera.view_projection_matrix();
+        let mut best_idx: Option<usize> = None;
+        let mut best_dist_sq = f32::MAX;
+
+        for (i, agent) in self.agents.iter().enumerate() {
+            if !agent.body.body.alive {
+                continue;
+            }
+            let pos = agent.body.body.position;
+            let clip = vp * glam::Vec4::new(pos.x, pos.y, pos.z, 1.0);
+            if clip.w <= 0.0 {
+                continue;
+            } // behind camera
+            let sx = clip.x / clip.w;
+            let sy = clip.y / clip.w;
+            let d = (sx - ndc_x).powi(2) + (sy - ndc_y).powi(2);
+            if d < best_dist_sq {
+                best_dist_sq = d;
+                best_idx = Some(i);
+            }
+        }
+
+        // Only select if click is reasonably close (within ~10% of screen)
+        if let Some(idx) = best_idx {
+            if best_dist_sq < 0.05 {
+                self.selected_agent_idx = idx;
+                self.agents[idx].trail_dirty = true;
+                let a = &self.agents[idx];
+                println!("[SELECT] Agent {} (gen {})", a.id, a.generation);
+                self.hud_dirty = true;
+            }
+        }
+    }
+
+    /// Build HUD overlay bars for the selected agent.
+    pub(crate) fn build_hud_bars(&self) -> Vec<HudBar> {
+        let Some(agent) = self.agents.get(self.selected_agent_idx) else {
+            return Vec::new();
+        };
+
+        let energy = agent.body.body.internal.energy_signal();
+        let integrity = agent.body.body.internal.integrity_signal();
+        let pred_err = agent.cached_prediction_error;
+        let explore = agent.cached_exploration_rate;
+
+        let bar_w = 0.35;
+        let bar_h = 0.025;
+        let left = -0.98;
+        let top = 0.97;
+        let gap = 0.035;
+        let bg = [0.15, 0.15, 0.15];
+
+        vec![
+            // Energy — green
+            HudBar {
+                x: left,
+                y: top,
+                width: bar_w,
+                height: bar_h,
+                fill: energy,
+                color: [0.2, 0.85, 0.2],
+                bg_color: bg,
+            },
+            // Integrity — blue
+            HudBar {
+                x: left,
+                y: top - gap,
+                width: bar_w,
+                height: bar_h,
+                fill: integrity,
+                color: [0.3, 0.5, 1.0],
+                bg_color: bg,
+            },
+            // Prediction error — red
+            HudBar {
+                x: left,
+                y: top - gap * 2.0,
+                width: bar_w,
+                height: bar_h,
+                fill: pred_err,
+                color: [0.95, 0.2, 0.15],
+                bg_color: bg,
+            },
+            // Exploration rate — yellow
+            HudBar {
+                x: left,
+                y: top - gap * 3.0,
+                width: bar_w,
+                height: bar_h,
+                fill: explore,
+                color: [0.95, 0.85, 0.1],
+                bg_color: bg,
+            },
+        ]
+    }
+}
